Create githubapi.ph
This script provides three functions:

get_repo(owner, repo) sends a GET request to the /repos/{owner}/{repo} endpoint to get information about a repository.
list_issues(owner, repo) sends a GET request to the /repos/{owner}/{repo}/issues endpoint to list issues on a repository.
create_issue(owner, repo, title, body) sends a POST request to the /repos/{owner}/{repo}/issues endpoint to create a new issue. The issue's title and body are passed as JSON in the body of the request.
Please remember to replace 'your_token_here', 'owner', and 'repo' with your actual GitHub access token, the owner of the repository, and the name of the repository, respectively.